Updated command with information about the Troika transport card

+++ src/troika_interaction/payments_troika.py
@@ -4,10 +4,12 @@ import os
 from aiogram import Router
 from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import StatesGroup, State
-from aiogram.types import CallbackQuery, Message, InlineKeyboardMarkup, InlineKeyboardButton, InputMediaPhoto, URLInputFile
+from aiogram.types import CallbackQuery, Message, InlineKeyboardMarkup, InlineKeyboardButton, InputMediaPhoto, \
+    URLInputFile
 import requests
 
 from src.models.mosmetro.search_response_troika import Troika
+from src.requests.get_token_authorization import get_new_access_token
 from src.requests.get_transport_card_info import get_troika_info
 from src.utils.load_config import load_config
 
@@ -17,20 +19,42 @@ TOP_UP_AMOUNTS = [10, 50, 100, 250, 500, 1000]
 
 troika_pay = Router()
 
+
 class TroikaStates(StatesGroup):
     waiting_manual_sum = State()
+
 
 @troika_pay.callback_query(lambda c: c.data.startswith("topup_"))
 async def show_topup_options(callback_query: CallbackQuery):
     card_number = callback_query.data.split("_")[1]
     buttons = [
-        [InlineKeyboardButton(text=f"💵 | Пополнить на {amount} рублей", callback_data=f"choosepay_{card_number}_{amount}")]
+        [InlineKeyboardButton(text=f"💵 | Пополнить на {amount} рублей",
+                              callback_data=f"choosepay_{card_number}_{amount}_4414")]
         for amount in TOP_UP_AMOUNTS
     ]
     buttons.append([InlineKeyboardButton(text="💷 | Пополнить на свою сумму", callback_data=f"manualsum_{card_number}")])
     buttons.append([InlineKeyboardButton(text="⬅️ | Назад", callback_data=f"back_to_card_{card_number}")])
     keyboard = InlineKeyboardMarkup(inline_keyboard=buttons)
     await callback_query.message.edit_reply_markup(reply_markup=keyboard)
+
+
+@troika_pay.callback_query(lambda c: c.data.startswith("tariff_"))
+async def show_tariff_options(callback_query: CallbackQuery):
+    card_number = callback_query.data.split("_")[1]
+    response = requests.get(f"{os.getenv('BACKEND_URL')}{os.getenv('PORT')}/mosmetro/troika/card_number/{card_number}")
+    troika_data = Troika(**response.json())
+    products = troika_data.availableProducts
+
+    buttons = []
+    for product in products:
+        button_text = f"🧾 | {product.name} за {product.price} рублей"
+        buttons.append([InlineKeyboardButton(text=button_text,
+                                             callback_data=f"choosepay_{card_number}_{product.price}_{product.id}")])
+    buttons.append([InlineKeyboardButton(text="⬅️ | Назад", callback_data=f"back_to_card_{card_number}")])
+
+    keyboard = InlineKeyboardMarkup(inline_keyboard=buttons)
+    await callback_query.message.edit_reply_markup(reply_markup=keyboard)
+
 
 @troika_pay.callback_query(lambda c: c.data.startswith("manualsum_"))
 async def start_manual_input(callback_query: CallbackQuery, state: FSMContext):
@@ -51,45 +75,6 @@ async def start_manual_input(callback_query: CallbackQuery, state: FSMContext):
     )
     await callback_query.message.edit_reply_markup(reply_markup=keyboard)
 
-@troika_pay.message(TroikaStates.waiting_manual_sum)
-async def handle_manual_sum(message: Message, state: FSMContext):
-    data = await state.get_data()
-    card_number = data.get("card_number")
-    prompt_id = data.get("manual_prompt_id")
-    markup_id = data.get("manual_markup_id")
-    try:
-        sum_value = int(message.text.strip())
-    except Exception:
-        sum_value = None
-    if not sum_value or sum_value < 10 or sum_value > 5000:
-        warn = await message.answer("<b>💳🚫 Сумма пополнения должна быть от 10 до 5000 рублей. Введите пожалуйста корректное значение.</b>")
-        await asyncio.sleep(5)
-        try:
-            await warn.delete()
-            await message.delete()
-        except Exception:
-            pass
-        return
-    try:
-        await message.delete()
-        await message.bot.delete_message(message.chat.id, prompt_id)
-    except Exception:
-        pass
-    keyboard = InlineKeyboardMarkup(
-        inline_keyboard=[
-            [
-                InlineKeyboardButton(text="🏦 | Банковская карта", callback_data=f"pay_{card_number}_{sum_value}_bankCard"),
-                InlineKeyboardButton(text="💠 | Система Быстрых Платежей", callback_data=f"pay_{card_number}_{sum_value}_sbp")
-            ],
-            [InlineKeyboardButton(text="⬅️ | Назад", callback_data=f"topup_{card_number}")]
-        ]
-    )
-    await state.update_data(manual_sum=sum_value)
-    await message.bot.edit_message_reply_markup(
-        chat_id=message.chat.id,
-        message_id=markup_id,
-        reply_markup=keyboard
-    )
 
 @troika_pay.message(TroikaStates.waiting_manual_sum)
 async def handle_manual_sum(message: Message, state: FSMContext):
@@ -102,10 +87,12 @@ async def handle_manual_sum(message: Message, state: FSMContext):
     except Exception:
         sum_value = None
     if not sum_value or sum_value < 10 or sum_value > 5000:
-        warn = await message.answer("<b>💳🚫 Сумма пополнения должна быть от 10 до 5000 рублей. Введите пожалуйста корректное значение.</b>")
+        warn = await message.answer(
+            "<b>💳🚫 Сумма пополнения должна быть от 10 до 5000 рублей. Введите пожалуйста корректное значение.</b>")
         await asyncio.sleep(15)
         try:
             await warn.delete()
+            print("!")
             await message.delete()
         except Exception:
             pass
@@ -115,21 +102,26 @@ async def handle_manual_sum(message: Message, state: FSMContext):
         await message.bot.delete_message(message.chat.id, prompt_id)
     except Exception:
         pass
+
     keyboard = InlineKeyboardMarkup(
         inline_keyboard=[
             [
-                InlineKeyboardButton(text="🏦 | Банковская карта", callback_data=f"pay_{card_number}_{sum_value}_bankCard"),
-                InlineKeyboardButton(text="💠 | Система Быстрых Платежей", callback_data=f"pay_{card_number}_{sum_value}_sbp")
+                InlineKeyboardButton(text="🏦 | Банковская карта",
+                                     callback_data=f"pay_{card_number}_{sum_value}_4414_bankCard"),
+                InlineKeyboardButton(text="💠 | Система Быстрых Платежей",
+                                     callback_data=f"pay_{card_number}_{sum_value}_4414_sbp")
             ],
             [InlineKeyboardButton(text="⬅️ | Назад", callback_data=f"topup_{card_number}")]
         ]
     )
     await state.update_data(manual_sum=sum_value)
+    await state.set_state(None)
     await message.bot.edit_message_reply_markup(
         chat_id=message.chat.id,
         message_id=markup_id,
         reply_markup=keyboard
     )
+
 
 @troika_pay.callback_query(lambda c: c.data.startswith("cancelmanual_"))
 async def cancel_manual_input(callback_query: CallbackQuery, state: FSMContext):
@@ -140,25 +132,36 @@ async def cancel_manual_input(callback_query: CallbackQuery, state: FSMContext):
     except Exception:
         pass
     await show_topup_options(callback_query)
+    await state.set_state(None)
+
 
 @troika_pay.callback_query(lambda c: c.data.startswith("choosepay_"))
 async def choose_payment_type(callback_query: CallbackQuery):
-    _, card_number, payment_sum = callback_query.data.split("_")
+    _, card_number, payment_sum, product_id = callback_query.data.split("_")
     payment_sum = int(payment_sum)
+    product_id = int(product_id)
+    if product_id == 4414:
+        callback_button = f'topup_{card_number}'
+    else:
+        callback_button = f'tariff_{card_number}'
+
     keyboard = InlineKeyboardMarkup(
         inline_keyboard=[
             [
-                InlineKeyboardButton(text="🏦 | Банковская карта", callback_data=f"pay_{card_number}_{payment_sum}_bankCard"),
-                InlineKeyboardButton(text="💠 | Система Быстрых Платежей", callback_data=f"pay_{card_number}_{payment_sum}_sbp")
+                InlineKeyboardButton(text="🏦 | Банковская карта",
+                                     callback_data=f"pay_{card_number}_{payment_sum}_{product_id}_bankCard"),
+                InlineKeyboardButton(text="💠 | Система Быстрых Платежей",
+                                     callback_data=f"pay_{card_number}_{payment_sum}_{product_id}_sbp")
             ],
-            [InlineKeyboardButton(text="⬅️ | Назад", callback_data=f"topup_{card_number}")]
+            [InlineKeyboardButton(text="⬅️ | Назад", callback_data=callback_button)]
         ]
     )
     await callback_query.message.edit_reply_markup(reply_markup=keyboard)
 
+
 @troika_pay.callback_query(lambda c: c.data.startswith("pay_"))
 async def process_payment(callback_query: CallbackQuery, state: FSMContext):
-    _, card_number, payment_sum, payment_type = callback_query.data.split("_")
+    _, card_number, payment_sum, product_id, payment_type = callback_query.data.split("_")
     payment_sum = int(payment_sum)
 
     response = requests.get(f"{os.getenv('BACKEND_URL')}{os.getenv('PORT')}/mosmetro/troika/card_number/{card_number}")
@@ -166,20 +169,31 @@ async def process_payment(callback_query: CallbackQuery, state: FSMContext):
         return
     troika_data = Troika(**response.json())
     card_number_data = troika_data.card.cardNumber
-    linked_card_id = 105147657
+    linked_card_id = troika_data.card.uid
+
+    data = await state.get_data()
+    refresh_token = data.get("refresh_token")
+    if refresh_token:
+        access_token, refresh_token = get_new_access_token(refresh_token)
+        await state.update_data(
+            access_token=access_token,
+            refresh_token=refresh_token
+        )
+    else:
+        access_token = None
 
     payload = {
         "callbackUrl": "mosmetro://redirect/payment",
-        "linkedCardId": linked_card_id,
+        "cardUid": linked_card_id,
         "paymentSum": payment_sum,
         "paymentType": payment_type,
         "saleType": "prepaid",
-        "ticketId": "4414"
+        "ticketId": product_id
     }
-
+    print(payload)
     headers = {
         "User-Agent": "MosMetro/4.2.3 (7874) (Android; samsung SM-A155F; 15; 2629830780)",
-        "Authorization": f"Bearer 2005A8A439B01F0139E3754AEEFBE0F52587ACFB474866DDF6E358B9372069E4"
+        "Authorization": f"Bearer {access_token}"
     }
 
     payment_response = requests.post("https://lk.mosmetro.ru/api/payments/v1.0", json=payload, headers=headers)
@@ -191,7 +205,8 @@ async def process_payment(callback_query: CallbackQuery, state: FSMContext):
             pay_keyboard = InlineKeyboardMarkup(
                 inline_keyboard=[
                     [InlineKeyboardButton(text=f"💱 | Оплатить {payment_sum} рублей", url=authorize_url)],
-                    [InlineKeyboardButton(text="🚫 | Отменить платёжную операцию", callback_data=f"back_to_card_{card_number}")]
+                    [InlineKeyboardButton(text="🚫 | Отменить платёжную операцию",
+                                          callback_data=f"back_to_card_{card_number}")]
                 ],
             )
             await callback_query.message.edit_reply_markup(reply_markup=pay_keyboard)
@@ -199,45 +214,45 @@ async def process_payment(callback_query: CallbackQuery, state: FSMContext):
             while True:
                 await asyncio.sleep(15)
                 check_response = requests.get(f"https://lk.mosmetro.ru/api/payments/v1.0/{session_id}", headers=headers)
+
+                keyboard = InlineKeyboardMarkup(
+                    inline_keyboard=[
+                        [InlineKeyboardButton(text="💸 | Пополнить",
+                                              callback_data=f"topup_{card_number}"),
+                         InlineKeyboardButton(text="🎟 | Тарифы",
+                                              callback_data=f"tariff_{card_number}")],
+                        [InlineKeyboardButton(text="🗑 | Удалить информацию", callback_data="delete_message")],
+                    ]
+                )
+
                 if check_response.ok:
                     check_data = check_response.json().get("data", {})
                     status = check_data.get("status")
                     if status == "success":
                         await callback_query.message.answer(
-                            f"💵✅ <b>Транспортная карта «Тройка» {card_number_data} была успешно пополнена на сумму {payment_sum} рублей! "
-                            f"Чтобы активировать пополнение, воспользуйтесь турникетом в «Московском Метрополитене», "
+                            f"💳✅ <b>Транспортная карта «Тройка» {card_number_data} получила платёж на сумму {payment_sum} рублей. "
+                            f"Чтобы активировать пополнение или тариф, воспользуйтесь турникетом в «Московском Метрополитене», "
                             f"валидатором в наземном транспорте, а также автоматом для продажи билетов.</b>",
                             show_alert=True)
-                        keyboard = InlineKeyboardMarkup(
-                            inline_keyboard=[
-                                [InlineKeyboardButton(text="💰 | Пополнить баланс",
-                                                      callback_data=f"topup_{card_number}")],
-                                [InlineKeyboardButton(text="🗑 | Удалить информацию", callback_data="delete_message")],
-                            ]
-                        )
                         await callback_query.message.edit_reply_markup(reply_markup=keyboard)
                         break
                     if status == "failed":
                         await callback_query.message.answer(
                             f"💵🚫 <b>Транспортная карта «Тройка» {card_number_data} не получила платёж на сумму {payment_sum} рублей. "
-                            f"Пополнение на данную транспортную карту не осуществлен. </b>",
+                            f"Пополнение или приобретение тарифа на данную транспортную карту не осуществлен. </b>",
                             show_alert=True)
-                        keyboard = InlineKeyboardMarkup(
-                            inline_keyboard=[
-                                [InlineKeyboardButton(text="💰 | Пополнить баланс",
-                                                      callback_data=f"topup_{card_number}")],
-                                [InlineKeyboardButton(text="🗑 | Удалить информацию", callback_data="delete_message")],
-                            ]
-                        )
                         await callback_query.message.edit_reply_markup(reply_markup=keyboard)
                         break
     else:
-        error_msg = await callback_query.message.answer(
-            f"💱🚫 <b>На текущий момент, способ оплаты неактивен и не отвечает с стороны сервера. "
-            f"Пожалуйста попробуйте позднее произвести данную операцию.</b>",
-            show_alert=True)
-        await asyncio.sleep(15)
-        await error_msg.delete()
+        if not refresh_token:
+            error_msg = await callback_query.message.answer(
+                f"💱🚫 <b>На текущий момент, способ оплаты неактивен и не отвечает с стороны сервера. "
+                f"Пожалуйста попробуйте позднее произвести данную операцию.</b>",
+                show_alert=True)
+            await asyncio.sleep(15)
+            await error_msg.delete()
+        # else:
+        #     get_new_access_token(refresh_token)
 
 
 # @troika_pay.callback_query(lambda c: c.data.startswith("checkpay_"))
@@ -288,7 +303,10 @@ async def back_to_card_view(callback_query: CallbackQuery):
     photo = URLInputFile(img, filename='troika.png')
     keyboard = InlineKeyboardMarkup(
         inline_keyboard=[
-            [InlineKeyboardButton(text="💰 | Пополнить баланс", callback_data=f"topup_{card_number}")],
+            [InlineKeyboardButton(text="💸 | Пополнить",
+                                  callback_data=f"topup_{card_number}"),
+             InlineKeyboardButton(text="🎟 | Тарифы",
+                                  callback_data=f"tariff_{card_number}")],
             [InlineKeyboardButton(text="🗑 | Удалить информацию", callback_data="delete_message")],
         ]
     )
